issue 2242 refactor structure of release notes agent
